TextDetection/requirements.txt
Flask==2.3.2
firebase-admin==6.0.1
pyrebase4==4.5.0
EasyOCR==1.6.2
opencv-python-headless==4.8.0.74
Pillow==9.4.0
